+++ zarrs/src/array/codec/array_to_bytes/packbits/packbits_partial_decoder.rs
@@ -53,10 +53,8 @@ fn partial_decode<'a>(
     let last_bit = last_bit.unwrap_or(component_size_bits - 1);
 
     // Get the component and element size in bits
-    let num_elements = decoded_representation.num_elements();
     let component_size_bits_extracted = last_bit - first_bit + 1;
     let element_size_bits = component_size_bits_extracted * num_components;
-    let elements_size_bytes = (num_elements * element_size_bits).div_ceil(8);
 
     let data_type_size_dec = decoded_representation
         .data_type()
@@ -66,7 +64,6 @@ fn partial_decode<'a>(
         })?;
 
     let element_size_bits_usize = usize::try_from(element_size_bits).unwrap();
-    let encoded_length_bits = elements_size_bytes * 8;
 
     let offset = match padding_encoding {
         PackBitsPaddingEncoding::FirstByte => 1,
@@ -76,13 +73,13 @@ fn partial_decode<'a>(
     let chunk_shape = decoded_representation.shape_u64();
     // Get the bit ranges that map to the elements
     let bit_ranges = indexer
-        .byte_ranges(&chunk_shape, element_size_bits_usize)?
+        .iter_contiguous_byte_ranges(&chunk_shape, element_size_bits_usize)?
         .collect::<Vec<_>>();
 
     // Convert to byte ranges, skipping the padding encoding byte
     let byte_ranges = bit_ranges.iter().map(|bit_range| {
-        let byte_start = offset + bit_range.start(encoded_length_bits).div(8);
-        let byte_end = offset + bit_range.end(encoded_length_bits).div_ceil(8);
+        let byte_start = offset + bit_range.start.div(8);
+        let byte_end = offset + bit_range.end.div_ceil(8);
         ByteRange::new(byte_start..byte_end)
     });
 
@@ -105,8 +102,8 @@ fn partial_decode<'a>(
         let mut component_idx_outer = 0;
         for (packed_elements, bit_range) in encoded_bytes.into_iter().zip(&bit_ranges) {
             // Get the bit range within the entire chunk
-            let bit_start = bit_range.start(encoded_length_bits);
-            let bit_end = bit_range.end(encoded_length_bits);
+            let bit_start = bit_range.start;
+            let bit_end = bit_range.end;
             let num_elements = (bit_end - bit_start) / element_size_bits;
 
             // Get the offset from the start of the byte range encapsulating the bit range