+++ zarrs/src/array/codec/bytes_to_bytes/zlib.rs
@@ -69,6 +69,8 @@ pub(crate) fn create_codec_zlib(metadata: &MetadataV3) -> Result<Codec, PluginCr
 mod tests {
     use std::{borrow::Cow, sync::Arc};
 
+    use zarrs_storage::byte_range::ByteRange;
+
     use crate::{
         array::{
             codec::{BytesPartialDecoderTraits, BytesToBytesCodecTraits, CodecOptions},
@@ -125,8 +127,9 @@ mod tests {
             .encode(Cow::Owned(bytes), &CodecOptions::default())
             .unwrap();
         let decoded_regions = ArraySubset::new_with_ranges(&[0..2, 1..2, 0..1])
-            .byte_ranges(array_representation.shape(), data_type_size)
-            .unwrap();
+            .iter_contiguous_byte_ranges(array_representation.shape(), data_type_size)
+            .unwrap()
+            .map(ByteRange::new);
         let input_handle = Arc::new(encoded);
         let partial_decoder = codec
             .partial_decoder(
@@ -173,8 +176,9 @@ mod tests {
             .encode(Cow::Owned(bytes), &CodecOptions::default())
             .unwrap();
         let decoded_regions = ArraySubset::new_with_ranges(&[0..2, 1..2, 0..1])
-            .byte_ranges(array_representation.shape(), data_type_size)
-            .unwrap();
+            .iter_contiguous_byte_ranges(array_representation.shape(), data_type_size)
+            .unwrap()
+            .map(ByteRange::new);
         let input_handle = Arc::new(encoded);
         let partial_decoder = codec
             .async_partial_decoder(